[ios][1/3] Fix Swift 6 errors - Fix non-final `Sendable` Exceptions (#35288)
# Why

We need to fix warnings that will become errors in the future. This is
the first of three PR's.
This one fixes this warning:
```
Non-final class 'PictureInPictureUnsupportedException' cannot conform to 'Sendable'; use '@unchecked Sendable'; this is an error in the Swift 6 language mode
```

# How

Makes all of our exceptions final so that they can conform to `Sendable`
without being marked as `@unchecked`

# Test Plan

Tested in BareExpo on iOS. The project compiles without the Swift 6
`Sendable` warning.